[core] Allows to add custom export item
